requirements.txt
streamlit
pyshark
openai
pandas>=2.2.0
plotly
python-dotenv
numpy
apscheduler